+++ jdbc/src/main/java/org/dbtools/android/room/jdbc/JdbcSQLiteOpenHelperFactory.kt
@@ -5,10 +5,12 @@ import androidx.sqlite.db.SupportSQLiteOpenHelper
 
 @Suppress("unused")
 class JdbcSQLiteOpenHelperFactory(
-        val path: String = ""
+        val path: String = "",
+        private val password: String = "",
+        private val onDatabaseConfigureBlock: (sqliteDatabase: JdbcSqliteDatabase) -> Unit = {}
 ) : SupportSQLiteOpenHelper.Factory {
 
     override fun create(configuration: SupportSQLiteOpenHelper.Configuration): SupportSQLiteOpenHelper {
-        return JdbcSQLiteOpenHelper(path, configuration.name, configuration.callback)
+        return JdbcSQLiteOpenHelper(path, configuration.name, configuration.callback, password, onDatabaseConfigureBlock)
     }
 }