+++ sqliteorg/src/main/java/org/dbtools/android/room/sqliteorg/SqliteOrgDatabaseUtil.kt
@@ -86,13 +86,16 @@ object SqliteOrgDatabaseUtil {
         var columnExists = false
 
         database.rawQuery("PRAGMA table_info($tableName)", null).use { cursor ->
-            cursor.moveToFirst()
-            do {
-                val currentColumn = cursor.getString(cursor.getColumnIndexOrThrow("name"))
-                if (currentColumn == columnName) {
-                    columnExists = true
-                }
-            } while (!columnExists && cursor.moveToNext())
+            if (cursor.moveToFirst()) {
+                do {
+                    val currentColumn = cursor.getString(cursor.getColumnIndexOrThrow("name"))
+                    if (currentColumn == columnName) {
+                        columnExists = true
+                    }
+                } while (!columnExists && cursor.moveToNext())
+            } else {
+                Timber.w("Query: [PRAGMA table_info($tableName)] returned NO data")
+            }
 
         }